Checking in all of the photos (Batch 2)
These are directly mirrored from the National Archives-maintained Flickr account, and are stored here simply for convenience in case they move in the future.

Original URL: https://www.flickr.com/photos/obamawhitehouse/

These images are produced by employees of the US Government and as such are largely in the public domain. The usage in this repository/project is noncommercial and does not imply or suggest the approval of the President, the First Family, or the White House.