Adding the unprocessed "memories" database
These come from the Flickr API for each of the photos. They have been very lightly cleaned up, but are still not in a format we can use for direct posting to services. Mostly they are a useful reference in case the Flickr API is restricted in the future or this data is lost.