Fixing photo credits
Some of the automated photographer extractions I'd done before didn't properly work on a few photos; this fixes them.

I also removed a few photos that didn't seem to belong (nomination photos for Justice Sotomayor) and others that were too small to be usable. Some photos in the Flickr album are super tiny but they are the only copy in there still.